backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
pytesseract==0.3.10
Pillow==11.3.0
Werkzeug==2.3.0
python-dotenv==1.0.0
